Refactoring: disbanded nanithing-dto module, DTOs go to interfaces, mappers to war
