+++ nanithing-jaxrs-interfaces/src/main/java/nanithing/web/jaxrs/GameDTO.java
@@ -1,4 +1,4 @@
-package nanithing.dto;
+package nanithing.web.jaxrs;
 
 import nanithing.model.GameState;
 import nanithing.model.GameVisibility;
