+++ nanithing-web/src/main/java/nanithing/web/GreetingResourceImpl.java
@@ -8,10 +8,9 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 import javax.inject.Inject;
 
-import nanithing.dto.GameDTO;
-import nanithing.dto.GameMapper;
+import nanithing.web.jaxrs.GameDTO;
+import nanithing.web.mappers.GameMapper;
 
-import nanithing.model.Game;
 import nanithing.model.ImmutableGame;
 
 /**
@@ -21,19 +20,6 @@ public class GreetingResourceImpl implements GreetingResource {
 
 	@Inject
 	private GameMapper gameMapper;
-
-	@Override
-	public CompletionStage<Game> hello() {
-		return CompletableFuture.completedStage(ImmutableGame.builder()
-				.id(UUID.randomUUID().toString())
-				.numberOfPlayers(2)
-				.numberOfRounds(5)
-				.roundTime(60)
-				.state(BUILDING)
-				.visibility(PUBLIC)
-				.build()
-		);
-	}
 
 	@Override
 	public CompletionStage<GameDTO> helloWithDTO() {
