+++ nanithing-web/src/main/java/nanithing/web/mappers/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Mappers from domain objects to DTOs.
+ */
+package nanithing.web.mappers;
