+++ pom.xml
@@ -15,7 +15,6 @@
 		<module>nanithing-model</module>
 		<module>nanithing-service-interfaces</module>
 		<module>nanithing-web</module>
-		<module>nanithing-dto</module>
 		<module>nanithing-dao-interfaces</module>
 		<module>nanithing-services</module>
 		<module>nanithing-types</module>
@@ -107,6 +106,11 @@
 				<groupId>org.liquibase</groupId>
 				<artifactId>liquibase-core</artifactId>
 				<version>${version.liquibase}</version>
+			</dependency>
+			<dependency>
+				<groupId>org.mapstruct</groupId>
+				<artifactId>mapstruct-jdk8</artifactId>
+				<version>1.3.1.Final</version>
 			</dependency>
 			<dependency>
 				<groupId>org.slf4j</groupId>
